requirements.txt
asgiref==3.4.1
Django==4.0.1
djangorestframework==3.13.1
django-environ==0.8.1
gevent==21.12.0
greenlet==1.1.2
importlib-metadata==4.10.0
Markdown==3.3.6
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
mypy==0.931
mypy-extensions==0.4.3
pytest==6.2.5
pytest-django==4.5.2
